requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pytest
httpx
openai